chores: Aggregated matchers, updated file-names
