--- src/utils/matchers/gh-username-matcher.ts
@@ -1,4 +1,0 @@
-export const getGithubUsername = (url: string) => {
-  const match = url.match(/github\.com\/([^/]+)/);
-  return match && match[1];
-};
