--- src/utils/matchers/tw-username-matcher.ts
@@ -1,4 +1,0 @@
-export const getTwitterUsername = (url: string) => {
-    const match = url.match(/(?:https?:\/\/)?(?:www\.)?twitter\.com\/(?:#!\/)?@?([^\/\?\s]*)/);
-    return match ? match[1] : undefined;
-}
